Solved single number II with bit manipulation approach
Since there are 3n + 1 numbers, we can't solve it with only XOR. But
there is a trick way to find which bit is owned by this single number.
That is to use an array with size 32 and sum every bits of all numbers,
if one of the bit cann't mod 3, that means this bit is owned by single
number.